refactor: isolates tests into their own lib
