+++ libs/core/question-bank/src/functions/get-new-variant.ts
