+++ libs/core/question-bank/src/functions/get-number-of-available-questions.ts
