+++ libs/core/question-bank/src/functions/get-question-preview.ts
