+++ libs/core/question-bank/src/index.ts
@@ -14,10 +14,6 @@ export * from "./entities/question-bank-question-true-or-false";
 export * from "./entities/question-bank-name";
 export * from "./entities/question-bank-subject";
 export * from "./entities/question-bank-validation";
-export * from "./entities/test-types";
-export * from "./questions/get-new-variant";
-export * from "./questions/get-question-preview";
-export * from "./tests/create-test";
-export * from "./tests/get-number-of-available-questions";
-export * from "./tests/create-test-question";
-export * from "./tests/process-test";
+export * from "./functions/get-new-variant";
+export * from "./functions/get-question-preview";
+export * from "./functions/get-number-of-available-questions";
