+++ libs/core/tests/.eslintrc.json
@@ -1,0 +1,5 @@
+{
+  "extends": ["../../../.eslintrc.json"],
+  "ignorePatterns": ["!**/*"],
+  "overrides": []
+}
