+++ libs/core/tests/src/entities/ids.ts
@@ -1,0 +1,2 @@
+export type TestId = string;
+export type QuestionId = string;
