+++ libs/core/tests/src/functions/create-test.ts
@@ -4,10 +4,10 @@ import {
   getRandomIdGenerator,
   getRandomShuffler,
 } from "@chair-flight/base/utils";
-import { questionBankNameSchema } from "../entities/question-bank-name";
+import { questionBankNameSchema } from "@chair-flight/core/question-bank";
 import { createTestQuestion } from "./create-test-question";
-import type { QuestionTemplate } from "../entities/question-bank-question";
-import type { Test } from "../entities/test-types";
+import type { Test } from "../entities/test";
+import type { QuestionTemplate } from "@chair-flight/core/question-bank";
 
 export const newTestConfigurationSchema = z.object({
   mode: z.enum(["study", "exam"]),
