+++ libs/core/tests/src/index.ts
@@ -1,0 +1,6 @@
+export * from "./entities/ids";
+export * from "./entities/test";
+export * from "./entities/test-question";
+export * from "./functions/create-test";
+export * from "./functions/create-test-question";
+export * from "./functions/process-test";
