+++ libs/core/tests/tsconfig.json
@@ -1,0 +1,8 @@
+{
+  "extends": "../../../tsconfig.base.json",
+  "include": ["src/**/*.ts"],
+  "files": ["../../../types.d.ts"],
+  "compilerOptions": {
+    "types": ["vitest/globals"]
+  }
+}
