+++ tsconfig.base.json
@@ -34,6 +34,7 @@
       "@chair-flight/core/blog": ["libs/core/blog/src/index.ts"],
       "@chair-flight/core/github": ["libs/core/github/src/index.ts"],
       "@chair-flight/core/question-bank": ["libs/core/question-bank/src/index.ts"],
+      "@chair-flight/core/tests": ["libs/core/tests/src/index.ts"],
       "@chair-flight/core/search": ["libs/core/search/src/index.ts"],
       "@chair-flight/providers/analytics": ["libs/providers/analytics/src/index.ts"],
       "@chair-flight/providers/blog": ["libs/providers/blog/src/index.ts"],
